Update translation.json
I've made some modifications to the Persian translation of the app to make it more appealing to the Persian users. Now it should be easier to understand and navigate with the app in the Persian version. I also made sure to be as grammatically correct as possible.
In Persian typing, there is a special non-printing character named "Zero-width non-joiner", which I replaced with the incorrect "Space character" uses in the previous translation whenever I saw one.
The translation is not perfect and there is still room for improvement, but I wanted to commit to this project. Since I lack programming knowledge, I did it this way.
Keep up with the good work.